Tehtävä C3.3 valmis

+++ bookstore/src/main/java/fi/roosakivila/bookstore/domain/Book.java
@@ -63,7 +63,7 @@ public class Book {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
